Setting up Push Notifications
Setting up push notifications for alerts for events and messages.  When a user signs up, the device token is sent to meteor.  Using actionable notifications so that the user can respond to event invitations and event messags without having the app open.  Created static method in InvitationActivityViewController so that the app can load up the event associated with the notification.