putting accessor to account property back in podfile
The swift account accessor method that I tried in the extension is not reliably accessing the property every time.  Moving the accessor out of METDDPClient+AccountsPassword.swift extension and putting it back in the METDDPClient.h file for now.  Also sending device token everytime the app sets it.  Before it was only being sent on signUp from app.  This caused issues if user signed up on web.