Moving Meteor pod and dependency pods into project
I have added custom code to Meteor-ios pod and the PocketSocket pod.  Taking the pods out of the podfile and moving the customized versions into the local project.  Changed _pmEnabled from YES to NO in line 113 in PSWebSocketDriver.m to disable compression in PocketSocket.  Meteor 1.2 and greater requires this fix.