Removing right side view controller to simplify tabBarController
TabBarController is not allowed to be the childviewcontroller of another parentviewcontroller.  Apple has human interface guidelines with rules about “More” on the TabBarController.  The more button should be used when there are 5 or more tabs.  The more button has a specific ViewController that the TabBarController manages.  I previously had the more button revealing the right side bar menu, but this is not the right way per Apple and I have updated accordingly.